Start minimized & On minimized
Start minimized
Start minimized will now work properly as the program is immediately
minimized right after it is started. The option in the Preference Dialog
Box is also re-enabled.

On minimized
The program now reads from the configuration file to decide if to "hide"
or to "minimize to taskbar". It will also write to the configuration
file according to the user's option in the preference dialog.