Created a Safe Zone for the AlarmModel
Created a Safe Zone for the AlarmModel. The safe zone is AlarmModelSafe.
Functionality is the same as  the TimerModelSafe safe zone but hours
reset to 0 after you incriment passed 23. This is because the alarm time
uses a 24 hour clock for comparison to the actual time. Also I fixed the
issues with the unit testing not finding private variables so now all
tests return true.